add platforms package with TelegramPlatform and WebSocketPlatform
remove use_ui from Bot.run, since it is a WebSocketPlatform specific parameter. Moved StreamlitUI file to platforms.websocket package

Bot's reset and receive_message receive the session_id as a parameter instead of the session object

add use_websocket and use_telegram methods to Bot. They create and return the websocket and telegram platforms

Session now has a reference to the platform it is connected to

replies are now done from the platforms and not from the session (to support platform-specific replies, e.g. audio or image replies), though we keep a reply method in the session for multi-platform replies support (only text replies as it must be a generic type of reply that all platforms support)

add default_body to avoid exceptions where a state body is (accidentally) not defined

add PlatformMismatchError for when trying to reply through a platform from a session with a different platform (this could happen if a bot has more than 1 active platforms)