refactor NLP module
Remove NLPConfiguration and merge all properties into Bot._config. There is a section 'nlp' for all properties related to NLP. To access any NLP property, there is a new wrapper method NLPEngine.get_property that accesses Bot._config's nlp section

Specific properties of SimpleIntentClassifier are now part of it as attributes.

Entity and intent processing methods are now part of entity and intent classes.

SimpleNER's training now performs the entity entries processing (before, it was done in NLPEngine.train, and it is not the proper place to do it)