add Property class
it is used to define bot properties. This is useful to determine different attributes of a property, like its type (e.g. int) or default value.

The preexisting bot properties definitions are now stored in their respective packages __init__.py files.

Add a wrapper to get NLP properties in NLPEngine