new Bot.stop() function to stop the bot (and its running platforms)
add 2 new boolean parameters to Bot.start(): sleep (used to not return from Bot.run()) and train (to choose weather to train the bot or not, maybe it is already trained)

refactor Bot._train() to Bot.train(), publicly available now

after the execution of Bot._init_global_states, Bot.global_initial_states is cleared, so when the bot is re-trained, this process is not duplicated

add BotNotTrainedError exception

refactor the Platform class: now it has the functions initialize, start, run and stop. The run function calls initialize and start. The stop function is called by the bot in Bot.stop()

new boolean attribute in Platform: running

update platforms wiki docs