fix telegram platform bug on unix systems
The class telegram.ext.Application, warns about this:

```
Not every :class:`asyncio.AbstractEventLoop` implements
:meth:`asyncio.loop.add_signal_handler`. Most notably, the standard event loop
on Windows, :class:`asyncio.ProactorEventLoop`, does not implement this method.
If this method is not available, stop signals can not be set.
```

In Unix systems stop signals are allowed, but since the telegram platform is not running in the main thread, this error happens in the telegram.ext.Application, when calling run_polling():

```
File "/Users/mgv99/Git/BESSER-Bot-Framework/venv/lib/python3.11/site-packages/telegram/ext/_application.py", line 946, in __run
    loop.add_signal_handler(sig, self._raise_system_exit)
  File "/Library/Frameworks/Python.framework/Versions/3.11/lib/python3.11/asyncio/unix_events.py", line 107, in add_signal_handler
    raise RuntimeError(str(exc))
RuntimeError: set_wakeup_fd only works in main thread of the main interpreter
```

To fix it, we now disable stop signals in the telegram application