update session flags
now there is a dictionary session.flags containing more than 1 flag, replacing file_flag. Right now, we need another flag for session.predicted_intent, for the same purpose as the session.file. It may happen that a transition different from the one associated with this intent is triggered (e.g. one evaluating some condition). In the following state, if there is a transition associated with the same intent, it should not be triggered as the time for this intent passed (unless the same intent is detected, in such case the flag will be set to true again).

session.flags contains 2 flags: 'predicted_intent' and 'file'. We may need more in the future

setting the file_flag to false has been moved from the file_received event to State.receive_file, since it was being set to false before checking all transitions. Similarly, the predicted_intent flag is set to false after checking all the transitions in State.receive_intent.