add MonitoringDB
This class is an interface to connect to a database where user interactions with the bot are stored to monitor the bot for later analysis.

Now, Bot has a new attribute _monitoring_db that, if enabled, on every user message creates a new thread to (in parallel) insert data into the database.

Included the new properties to set the db credentials.